document validated source
